[static-site-prereqs] feat: move getAllPatternsObj to actual file
