+++ packages/larva/__tests__/lib/utils/getAllPatternsObj.test.js
@@ -1,10 +1,6 @@
 const path = require( 'path' );
 const getAppConfiguration = require( '../../../lib/utils/getAppConfiguration' );
-const getPatternVariants = require( '../../../lib/utils/getPatternVariants' );
-const getSubDirectoryNames = require( '../../../lib/utils/getSubDirectoryNames' );
-const getModuleNamesFromDirectory = require( '../../../lib/utils/getModuleNamesFromDirectory' );
-
-const config = getAppConfiguration( 'patterns' );
+const getAllPatternsObj = require( '../../../lib/utils/getAllPatternsObj' );
 
 const expectedPatterns = {
 	'components': {
@@ -30,39 +26,14 @@ const expectedPatterns = {
 		]
 	}
 }
+const config = getAppConfiguration( 'patterns' );
 
 describe( 'getAllPatternsObj', () => {
 
 	it( 'returns an object containing all patterns and variants', () => {
-		function getAllPatterns( startPath ) {
-
-			const patternDirs = getSubDirectoryNames( startPath );
-
-			let obj = {};
-
-			patternDirs.map( ( patternType ) => {
-				let patterns = [];
-				obj[patternType] = {};
-
-				// Modules can be ignored, which maybe is unecessary
-				// but for now continue on that path.
-				if ( 'modules' === patternType ) {
-					patterns = getModuleNamesFromDirectory( startPath, config.ignoredModules );
-				} else {
-					patterns = getSubDirectoryNames( path.join( startPath, patternType ) );
-				}
-
-				patterns.map( ( pattern ) => {
-					obj[patternType][pattern] = getPatternVariants( path.join( startPath, `${patternType}/${pattern}` ) );
-				} );
-
-			} );
-
-			return obj;
-		}
 
 		expect( 
-			getAllPatterns( config.projectPatternsDir )
+			getAllPatternsObj( config.projectPatternsDir )
 		).toEqual( 
 			expectedPatterns
 		);
